+++ bundle/src/test/java/com/adobe/acs/commons/mcp/impl/processes/TagReportTest.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -39,6 +39,7 @@ import org.mockito.stubbing.Answer;
 import com.adobe.acs.commons.fam.ActionManager;
 import com.adobe.acs.commons.fam.actions.Actions;
 import com.adobe.acs.commons.functions.CheckedConsumer;
+import com.adobe.acs.commons.mcp.impl.processes.TagReporter.ReportColumns;
 import com.adobe.acs.commons.mcp.util.StringUtil;
 
 import io.wcm.testing.mock.aem.junit.AemContext;
@@ -124,7 +125,9 @@ public class TagReportTest {
     tagReporter.tagPath = "/etc/tags/workflow";
     tagReporter.rootSearchPath = "/content";
     tagReporter.includeReferences = true;
+    tagReporter.referencesCharacterLimit = "4096";
 
+    tagReporter.init();
     tagReporter.traverseTags(actionManager);
     tagReporter.recordTags(actionManager);
     assertEquals(11, tagReporter.getReportRows().size());
@@ -141,6 +144,28 @@ public class TagReportTest {
             StringUtils.isBlank((String) r.get(TagReporter.ReportColumns.REFERENCES)));
       }
     });
+  }
+
+  @Test
+  public void testLargeCell() throws Exception {
+
+    ctx.load().json("/com/adobe/acs/commons/mcp/impl/processes/lotsofchildren.json", "/content/lotsofchildren");
+
+    tagReporter.tagPath = "/etc/tags/workflow/wcm/translation";
+    tagReporter.rootSearchPath = "/content";
+    tagReporter.includeReferences = true;
+    tagReporter.referencesCharacterLimit = "4096";
+
+    tagReporter.init();
+    tagReporter.traverseTags(actionManager);
+    tagReporter.recordTags(actionManager);
+
+    assertEquals(1, tagReporter.getReportRows().size());
+
+    assertEquals(StringUtil.getFriendlyName(TagReporter.ItemStatus.SUCCESS.name()),
+        tagReporter.getReportRows().get(0).get(TagReporter.ReportColumns.STATUS));
+
+    assertEquals(4096, tagReporter.getReportRows().get(0).get(ReportColumns.REFERENCES).toString().length());
   }
 
 }